# Conflicts:
#	include/mmakefile.src
#	src/aros/mesa3dgl/compiler/mmakefile.src
#	src/aros/mesa3dgl/galliumaux/mmakefile.src
#	src/aros/mesa3dgl/galliumvm/mmakefile.src
#	src/aros/mesa3dgl/glapi/mmakefile.src
#	src/aros/mesa3dgl/mesa/mmakefile.src
#	src/aros/mesa3dgl/mesa3dgl_glacreatecontext.c
#	src/aros/mesa3dgl/mesautil/mmakefile.src
#	src/aros/mesa3dgl/mmakefile.src
#	src/gallium/auxiliary/util/u_format_s3tc.c
#	src/gallium/state_trackers/aros/hidds/gallium/gallium_class.c
#	src/gallium/state_trackers/aros/hidds/gallium/mmakefile.src
#	src/gallium/state_trackers/aros/hidds/intelgma/edid/edid.c
#	src/gallium/state_trackers/aros/hidds/intelgma/edid/mmakefile.src
#	src/gallium/state_trackers/aros/hidds/intelgma/i915/mmakefile.src
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_bitmapclass.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_copybox.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_display.h
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_displayclass.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_galliumclass.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_hidd.h
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_hiddclass.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_init.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_startup.c
#	src/gallium/state_trackers/aros/hidds/intelgma/intelgma_support.c
#	src/gallium/state_trackers/aros/hidds/intelgma/mmakefile.src
#	src/gallium/state_trackers/aros/hidds/softpipe/softpipe_galliumclass.c
#	src/gallium/state_trackers/aros/hidds/vc4/mmakefile.src
#	src/gallium/state_trackers/aros/hidds/vc4/vc4_intern.h
#	src/gallium/state_trackers/aros/hidds/vc4/vc4galliumclass.c
#	src/gallium/state_trackers/aros/hidds/vmwaresvga/mmakefile.src
#	src/gallium/state_trackers/aros/library/createpipe.c
#	src/gallium/state_trackers/aros/library/mmakefile.src

+++ src/compiler/glsl/lower_ubo_reference.cpp
@@ -463,7 +463,7 @@ lower_ubo_reference_visitor::ssbo_store(void *mem_ctx,
    call_params.push_tail(offset->clone(mem_ctx, NULL));
    call_params.push_tail(deref->clone(mem_ctx, NULL));
    call_params.push_tail(new(mem_ctx) ir_constant(write_mask));
-   call_params.push_tail(new(mem_ctx) ir_constant(ssbo_access_params()));
+   call_params.push_tail(new(mem_ctx) ir_constant((unsigned int)ssbo_access_params()));
    return new(mem_ctx) ir_call(sig, NULL, &call_params);
 }
 
@@ -504,7 +504,7 @@ lower_ubo_reference_visitor::ssbo_load(void *mem_ctx,
    exec_list call_params;
    call_params.push_tail(this->uniform_block->clone(mem_ctx, NULL));
    call_params.push_tail(offset->clone(mem_ctx, NULL));
-   call_params.push_tail(new(mem_ctx) ir_constant(ssbo_access_params()));
+   call_params.push_tail(new(mem_ctx) ir_constant((unsigned int)ssbo_access_params()));
 
    return new(mem_ctx) ir_call(sig, deref_result, &call_params);
 }
